refactor: Move participant components into core
refactor: Move participant components into core/

refactor: Move participant components into core/
